api: Rework the entire constructor logic for Device
This mainly deprecates the NewDevice constructor, which, as a matter of fact, is currently
pushing users towards the wrongest possible behaviours. Instead, create a new DeviceOptions
struct that should satisfy all advanced users' needs, and expose all possible configuration
options for Device. Ideally, NewDevice should be gone by 1.0.

NewDeviceWithPersistency should become, instead, the main and default constructor, which
currently provides the safest and soundest set of options for all users.

In this process, a number of bad behaviours in terms of creating directories and handling
variables have been fixed, which at the end of the day should be fairly beneficial in terms
of code clarity and performance (we're no longer stat-ing the filesystem every now and then).

Fixes #23

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>