store/api: Allow specifying the DB storage explicitly
On one hand, this allows users more control over what they want to use (in fact, this
opens the door to any gorm supported DB), on the other one it relieves the requirement
of using CGO at all times and even not using any storage at all. All of this without
changing the APIs.

The only behavior change (which doesn't matter much as it matches the behavior of the
SDK up until a couple days ago) is that when no persistency is specified, no databases
will be created.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>