Do not deadlock when disconnection corner case happens
If a device were to disconnect from the broker before having sent
its introspection, the connection lost handler would deadlock
trying to acquire the lock on the inflight message queue which is
still locked.
Fix this by moving the first lock to the connection handler,
and check that another lock attempt is performed when connection
is lost only if the device is not allowed to reconnect.
This still guarantees that the inflight message queue lock is
released only when (eventually) the introspection is sent.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>